# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Public Healthcare Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List all alerts
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a new alert (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertBody"
      responses:
        "201":
          description: Alert created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    put:
      operationId: updateAlert
      tags: [alerts]
      summary: Update an alert (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertBody"
      responses:
        "200":
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
    delete:
      operationId: deleteAlert
      tags: [alerts]
      summary: Delete an alert (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /vaccinations:
    get:
      operationId: listVaccinations
      tags: [vaccinations]
      summary: List vaccination info
      responses:
        "200":
          description: List of vaccination info
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vaccination"
    post:
      operationId: createVaccination
      tags: [vaccinations]
      summary: Create vaccination info (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVaccinationBody"
      responses:
        "201":
          description: Vaccination info created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vaccination"
  /vaccinations/{id}:
    put:
      operationId: updateVaccination
      tags: [vaccinations]
      summary: Update vaccination info (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVaccinationBody"
      responses:
        "200":
          description: Vaccination info updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vaccination"
    delete:
      operationId: deleteVaccination
      tags: [vaccinations]
      summary: Delete vaccination info (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vaccination info deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all user reports (admin only)
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: createReport
      tags: [reports]
      summary: Create a new self-report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
  /reports/{id}:
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /blogs:
    get:
      operationId: listBlogs
      tags: [blogs]
      summary: List all blogs
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: author
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of blogs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Blog"
    post:
      operationId: createBlog
      tags: [blogs]
      summary: Create a new blog post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogBody"
      responses:
        "201":
          description: Blog created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
  /blogs/{id}:
    get:
      operationId: getBlog
      tags: [blogs]
      summary: Get a single blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
        "404":
          description: Blog not found
    put:
      operationId: updateBlog
      tags: [blogs]
      summary: Update a blog post (author only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlogBody"
      responses:
        "200":
          description: Blog updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blog"
    delete:
      operationId: deleteBlog
      tags: [blogs]
      summary: Delete a blog post (author or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /blogs/{id}/like:
    post:
      operationId: toggleBlogLike
      tags: [blogs]
      summary: Toggle like on a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
  /blogs/{id}/comments:
    post:
      operationId: createBlogComment
      tags: [blogs]
      summary: Add a comment to a blog post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /blogs/{id}/comments/{commentId}:
    delete:
      operationId: deleteBlogComment
      tags: [blogs]
      summary: Delete a comment (author or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comment deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        location:
          type: string
      required:
        - name
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        location:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    Alert:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - message
        - severity
        - isActive
        - createdAt
        - updatedAt
    CreateAlertBody:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        isActive:
          type: boolean
      required:
        - title
        - message
        - severity
    Vaccination:
      type: object
      properties:
        id:
          type: integer
        disease:
          type: string
        vaccineName:
          type: string
        description:
          type: string
        dosesRequired:
          type: integer
        eligibility:
          type: string
        availability:
          type: string
        area:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - disease
        - vaccineName
        - description
        - dosesRequired
        - eligibility
        - availability
        - area
        - createdAt
        - updatedAt
    CreateVaccinationBody:
      type: object
      properties:
        disease:
          type: string
        vaccineName:
          type: string
        description:
          type: string
        dosesRequired:
          type: integer
        eligibility:
          type: string
        availability:
          type: string
        area:
          type: string
      required:
        - disease
        - vaccineName
        - description
        - dosesRequired
        - eligibility
        - availability
        - area
...
    Report:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        date:
          type: string
          format: date-time
        phoneNumber:
          type: string
        symptoms:
          type: string
        location:
          type: string
        additionalInfo:
          type: string
        userId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - date
        - phoneNumber
        - symptoms
        - location
        - userId
        - createdAt
    CreateReportBody:
      type: object
      properties:
        name:
          type: string
        date:
          type: string
          format: date-time
        phoneNumber:
          type: string
        symptoms:
          type: string
        location:
          type: string
        additionalInfo:
          type: string
      required:
        - name
        - date
        - phoneNumber
        - symptoms
        - location
    Blog:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        category:
          type: string
        description:
          type: string
        content:
          type: string
        author:
          type: string
        authorName:
          type: string
        likes:
          type: array
          items:
            type: string
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - category
        - description
        - author
        - authorName
        - likes
        - comments
        - createdAt
        - updatedAt
    CreateBlogBody:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        description:
          type: string
        content:
          type: string
      required:
        - title
        - category
        - description
    Comment:
      type: object
      properties:
        _id:
          type: string
        userId:
          type: string
        userName:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - _id
        - userId
        - userName
        - content
        - createdAt
    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    LikeResponse:
      type: object
      properties:
        likes:
          type: integer
        isLiked:
          type: boolean
      required:
        - likes
        - isLiked