# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Public Healthcare Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  _id: string;
  userId: string;
  userName: string;
  content: string;
  createdAt: Date;
}