# lib/api-zod/src/generated/types/createBlogBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Public Healthcare Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBlogBody {
  title: string;
  category: string;
  description: string;
  content?: string;
}